+++ benchmarks/run.sh
@@ -1,0 +1,2 @@
+julia runtests.jl 100000 50000000
+julia runtests.jl 5000000 10000000